Complete previous commit and make new `Types.hs`
I forgot to add the code to `ReplUi` from `Gameplay` :(

Make new `Types.hs` file to prevent cyclic module dependency between
`Gameplay` and `ReplUi` because they both need `BaseStateT` (which was
originally in the `Gameplay` module).

I also decided to move all the types there. Hopefully this is a pattern
that works (although I hear it might be an anti-pattern online; we'll
see if it is!).
